In progress. Still struggling with the mask issues
